app/api/v1/auth/requirements.txt
fastapi==0.99.0
mangum==0.11.0
pydantic==1.10.9
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.34.34
requests==2.31.0
uvicorn
email-validator==2.1.0 